# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: FinFlow API
  description: 家計・サブスク管理アプリ FinFlow のバックエンドAPI
  version: "1.0.0"
servers:
  - url: http://localhost:5000
    description: 開発環境
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        userId:
          type: string
        email:
          type: string
        expiresAt:
          type: string
          format: date-time

    # ---- Category ----
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        statusCode:
          type: integer
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: ユーザー登録
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: 登録成功・JWTトークン返却
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: ログイン
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ログイン成功・JWTトークン返却
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: 認証失敗

  # ======== Categories ========
...